+++ src/sys/win/selector.rs
@@ -58,13 +58,13 @@ impl Selector {
 
             if self.read_sets.fd_count > 0 {
                 for i in 0 .. self.read_sets.fd_count {
-                    evts.push(EventEntry::new_evfd(self.read_sets.fd_array[i as usize] as u64, FLAG_READ));
+                    evts.push(EventEntry::new_evfd(self.read_sets.fd_array[i as usize] as u32, FLAG_READ));
                 }
                 size += self.read_sets.fd_count;
             }
             if self.write_sets.fd_count > 0 {
                 for i in 0 .. self.write_sets.fd_count {
-                    evts.push(EventEntry::new_evfd(self.write_sets.fd_array[i as usize] as u64, FLAG_WRITE));
+                    evts.push(EventEntry::new_evfd(self.write_sets.fd_array[i as usize] as u32, FLAG_WRITE));
                 }
                 size += self.write_sets.fd_count;
             }
@@ -72,7 +72,7 @@ impl Selector {
         Ok(size)
     }
 
-    pub fn register(&mut self, fd : u64, ev_events : EventFlags) {
+    pub fn register(&mut self, fd : u32, ev_events : EventFlags) {
         let fd = fd as SOCKET;
         if ev_events.contains(FLAG_READ) && !self.read_sockets.contains(&fd) {
             self.read_sockets.push(fd);
@@ -82,7 +82,7 @@ impl Selector {
         }
     }
 
-    pub fn deregister(&mut self, fd : u64, _ : EventFlags) {
+    pub fn deregister(&mut self, fd : u32, _ : EventFlags) {
         let fd = fd as SOCKET;
         fn search_index(vec : &Vec<SOCKET>, value : &SOCKET) -> Option<usize> {
             for i in 0 .. vec.len() {
