+++ src/timer.rs
@@ -9,8 +9,8 @@ extern crate time;
 
 pub struct Timer {
     timer_queue: BinaryHeap<EventEntry>,
-    time_sets : HashSet<u64>,
-    time_id : u64,
+    time_sets : HashSet<u32>,
+    time_id : u32,
 }
 
 impl Timer {
@@ -27,7 +27,7 @@ impl Timer {
 	}
 
 	// ID = 0 为无效ID
-	pub fn add_timer(&mut self, mut entry : EventEntry) -> u64 {
+	pub fn add_timer(&mut self, mut entry : EventEntry) -> u32 {
 		if entry.ev_fd == 0 {
 			entry.ev_fd = self.calc_new_id();	
 		};
@@ -36,7 +36,7 @@ impl Timer {
 		self.time_id
 	}
 
-	pub fn del_timer(&mut self, time_id : u64) {
+	pub fn del_timer(&mut self, time_id : u32) {
 		let mut data = Vec::new();
 		while let Some(entry) = self.timer_queue.pop() {
 			if entry.ev_fd != time_id {
@@ -61,7 +61,7 @@ impl Timer {
 		self.timer_queue.pop()
 	}
 
-	fn calc_new_id(&mut self) -> u64 {
+	fn calc_new_id(&mut self) -> u32 {
 		loop {
 			self.time_id += 1;
 			self.time_id = cmp::max(self.time_id, 1);
